Managed all models through Core Data
The app could already fetch all the data necessary from the Internet and
display it, but eventually it should also be able to store the data
offline so that the Internet isn't needed for viewing recipes.

This is the first step in that direction. It adds the Core Data modeling
in place so that all of the models are now managed by Core Data.