refactor core data models
This refactors the app to separate the Core Data entities from the basic
models. This will allow the UI to not have to know about the fact that
it's interacting with Core Data, and should allow for a smoother sync
process without having to worry about side effects as much.