Make Core Data the source of truth
Rather than require users to hit the Ressipy API every time they load
the category list, a category, or a recipe, that will now all be fetched
by Core Data. The app will run a sync with the server in the background
whenever it's opened, but if this fails users will still be able to view
the categories and recipes that they already have downloaded.