add basic authentication
This adds the new AuthManager singleton and account screen. When a user
logs in, the app will now create a token and store it in the secure
enclave. The user's email address and a list of permissions returned by
the server are also stored in the user defaults.

This will allow users to log in, and with this in place, the app will be
able to add an Authorization header to every request so that the server
can attach most of the requests to authenticated users. This will
improve the effectiveness of rate limiting too. And it paves the way for
creating, updating, and deleting records from the app as well.